+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Gitaly changelog
 
 UNRELEASED
-- Send raw body on CommitService.CommitsBetween
+- Send parent ids and raw body on CommitService.CommitsBetween
   https://gitlab.com/gitlab-org/gitaly/merge_requests/216
 
 v0.18.0